+++ src/observer/storage/default/default_storage_stage.cpp
@@ -196,7 +196,7 @@ void DefaultStorageStage::handle_event(StageEvent *event) {
         case SCF_INSERT: { // insert into
             const Inserts &inserts = sql->sstr.insertion;
             const char *table_name = inserts.relation_name;
-            for (int i = 0; i < inserts.value_num; i++) {
+            for (int i = 0; i < (int)inserts.value_num; i++) {
                 if (inserts.values[i].type == DATES && !isValidDate(*((int *) inserts.values[i].data)))
                     rc = RC::SCHEMA_FIELD_TYPE_MISMATCH;
             }
@@ -213,7 +213,7 @@ void DefaultStorageStage::handle_event(StageEvent *event) {
             int updated_count = 0;
             if (updates.value.type == DATES && !isValidDate(*((int *) updates.value.data)))
                 rc = RC::SCHEMA_FIELD_TYPE_MISMATCH;
-            for (int i = 0; i < updates.condition_num; i++) {
+            for (int i = 0; i < (int)updates.condition_num; i++) {
                 if (!updates.conditions[i].left_is_attr &&
                     updates.conditions[i].left_value.type == DATES &&
                     !isValidDate(*((int *) updates.conditions[i].left_value.data))) {
@@ -238,7 +238,7 @@ void DefaultStorageStage::handle_event(StageEvent *event) {
             const Deletes &deletes = sql->sstr.deletion;
             const char *table_name = deletes.relation_name;
             int deleted_count = 0;
-            for (int i = 0; i < deletes.condition_num; i++) {
+            for (int i = 0; i < (int)deletes.condition_num; i++) {
                 if (!deletes.conditions[i].left_is_attr &&
                     deletes.conditions[i].left_value.type == DATES &&
                     !isValidDate(*((int *) deletes.conditions[i].left_value.data))) {
